move render command to pkg/cmd
