+++ pkg/cmd/render/testdata/auth/kubeconfig
