+++ pkg/cmd/render/testdata/configs/kube-controller-manager-config-overrides.yaml
