+++ pkg/cmd/render/testdata/tls/admin.key
