+++ pkg/cmd/render/testdata/tls/aggregator-ca.crt
