+++ pkg/cmd/render/testdata/tls/aggregator-ca.key
