+++ pkg/cmd/render/testdata/tls/apiserver-proxy.crt
