+++ pkg/cmd/render/testdata/tls/apiserver.key
