+++ pkg/cmd/render/testdata/tls/cluster-apiserver-ca.crt
