+++ pkg/cmd/render/testdata/tls/etcd-client-ca.crt
