+++ pkg/cmd/render/testdata/tls/etcd-client.crt
