+++ pkg/cmd/render/testdata/tls/kube-ca.key
