+++ pkg/cmd/render/testdata/tls/kubelet.crt
