+++ pkg/cmd/render/testdata/tls/machine-config-server.key
