+++ pkg/cmd/render/testdata/tls/openshift-apiserver.crt
