+++ pkg/cmd/render/testdata/tls/openshift-apiserver.key
