+++ pkg/cmd/render/testdata/tls/service-serving-ca.crt
